flask_app/requirements.txt
Flask
Werkzeug

ctranslate2

pydub
PyAudio
sounddevice

opencc-python-reimplemented

numpy
ffmpeg-python
huggingface_hub
tokenizers
tqdm